Optimized Water + Guard Stairs
Water works on a two mesh, two material setup. One smaller mesh uses the follower script to snap to a grid around the player (snapping to keep triangles in same relative location)--this mesh deforms based on wave height using a more complex water shader. The other mesh is a massive, two-triangle square that uses a simpler shader that does not have any vertex displacement. I use a simple distance fade to make these meshes blend together. This method only runs two shaders for the water (instead of one per chunk like before), and it also doesn't need to update LOD meshes in real time.

The guard is also back in working condition. They can walk up stairs now, and I also added RNG to their IDLE_POINT state (when they go inspect set points during their patrol). This state can be used to inspect sounds the player or world makes in the future and isn't tied to the RNG directly.

I started a simpler tail for Carmelita that will hopefully be brought over to Sly once it's finished. Right now it's a mess, so please ignore it :)

I also reworked a lot of the environment lights, colors, and the final saturation/contrast adjustment in the Environment node for Paris.